requirements.txt
torch==2.3.1 #pip install torch==2.3.1 torchvision==0.18.1 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.18.1
setuptools==69.5.1
gradio==4.38.1
openai>=1.0.0
opencv_python==4.8.1.78
diffusers[torch]==0.29.2
bitsandbytes
termcolor
fairscale
natsort
omegaconf
pycocotools
matplotlib
onnxruntime
onnx
timm
openpyxl
-e ./pretrained-models/sam-hq/
-e ./pretrained-models/GroundingDINO/
-e ./pretrained-models/recognize-anything/
transformers==4.42.4 # Somehow, we needed to put transformers after GroundingDINO.